refactor: Service ->api로 변경, Context 폴더 밖으로 빼기, 후원버튼 text 변경하기
