+++ src/pages/List/components/CreditModal/CreditModal.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useContext } from "react";
 
 import "./CreditModal.css";
-import { CreditContext } from "../../../../components/CreditContextProvider";
+import { CreditContext } from "../../../../context/CreditContextProvider";
 
 function CreditModal({ onClose }) {
   const { handleRecharge } = useContext(CreditContext);
