+++ src/pages/List/components/DonationsModal/DonationsModal.jsx
@@ -7,7 +7,7 @@ import useEscapeModal from "../../../../hooks/useEscapeModal";
 import useDonationHandler from "../../../../hooks/useDonationHandler";
 import handleBackgroundClick from "../../../../utils/handleBackgroundClick";
 
-import { CreditContext } from "../../../../components/CreditContextProvider";
+import { CreditContext } from "../../../../context/CreditContextProvider";
 import CloseButton from "./CloseButton";
 
 function DonationsModal({
@@ -39,7 +39,7 @@ function DonationsModal({
     localCredit,
     selectedDonation,
     updateProgressbar,
-    closeModal,
+    closeModal
   );
 
   useEscapeModal(closeModal);
